requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4

python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.5.1
pandas==2.2.2
PyMuPDF==1.24.7
docx2txt==0.8
google-api-python-client==2.142.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
tenacity==8.3.0
python-dateutil==2.9.0.post0
rich==13.7.1
pytest==8.3.2